# back/openapi '3.0.yml
openapi: '3.0.3'
info:
  title: スケジュール管理API
  version: '1.0.0'
  description: |
    勤怠スケジュールを管理・共有するバックエンドAPIの仕様書です。

servers:
  - url: http://localhost:8080
    description: ローカル開発サーバー

security:
  - sessionAuth: []

paths:
  /api/login:
    post:
      summary: ログイン
      description: |
        フォーム送信されたユーザーIDとパスワードで認証し、セッションを開始します。
        認証成功時は `JSESSIONID` クッキーがレスポンスに含まれます。
      operationId: login
      tags:
        - 認証
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 認証成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccess'
              example:
                message: Login successful
                user: "00001"
        '401':
          description: 認証に失敗しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
              example:
                message: Bad credentials

  /api/schedule:
    get:
      summary: スケジュール一覧取得
      description: 指定条件に合致するユーザーごとのスケジュールを取得します。
      operationId: getSchedule
      tags:
        - スケジュール
      parameters:
        - name: userId
          in: query
          required: false
          description: ユーザーID（最大5桁）
          schema:
            type: string
            pattern: '^[0-9０-９]{0,5}$'
            example: "00001"
        - name: week
          in: query
          required: false
          description: 対象週（YYYY-Www）。`viewMode` が `WEEK` の場合に必須。
          schema:
            type: string
            pattern: '^\\d{4}-W(0[1-9]|[1-4][0-9]|5[0-3])$'
            example: "2025-W02"
        - name: month
          in: query
          required: false
          description: 対象月（YYYY-MM）。`viewMode` が `MONTH` の場合に必須。
          schema:
            type: string
            pattern: '^\\d{4}-(0[1-9]|1[0-2])$'
            example: "2025-01"
        - name: viewMode
          in: query
          required: true
          description: 取得モード
          schema:
            type: string
            enum: [WEEK, MONTH]
            example: WEEK
        - name: name
          in: query
          required: false
          description: 氏名（部分一致、最大50文字）
          schema:
            type: string
            maxLength: 50
        - name: organizationCode
          in: query
          required: false
          description: 組織コード（最大3桁）
          schema:
            type: string
            pattern: '^\\d{0,3}$'
            example: "006"
      responses:
        '200':
          description: 正常に取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message:
                  - "period: Only one of week or month can be provided"
                timestamp: "2025-01-01T09:00:00"
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

  /api/irregularSchedule:
    get:
      summary: 非定期スケジュール取得
      operationId: getIrregularSchedule
      tags:
        - 非定期スケジュール
      parameters:
        - name: from
          in: query
          required: true
          description: 検索開始日（YYYY-MM-DD）
          schema:
            type: string
            format: date
            example: "2025-01-01"
        - name: to
          in: query
          required: true
          description: 検索終了日（YYYY-MM-DD）
          schema:
            type: string
            format: date
            example: "2025-01-31"
      responses:
        '200':
          description: 正常に取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IrregularScheduleDTO'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
    post:
      summary: 非定期スケジュール登録
      operationId: postIrregularSchedule
      tags:
        - 非定期スケジュール
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IrregularScheduleInsertForm'
      responses:
        '201':
          description: 登録に成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
    put:
      summary: 非定期スケジュール更新
      operationId: putIrregularSchedule
      tags:
        - 非定期スケジュール
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IrregularScheduleUpdateForm'
      responses:
        '204':
          description: 更新に成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

  /api/irregularSchedule/{id}:
    delete:
      summary: 非定期スケジュール削除
      operationId: deleteIrregularSchedule
      tags:
        - 非定期スケジュール
      parameters:
        - name: id
          in: path
          required: true
          description: 削除対象スケジュールID
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: 削除に成功
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

  /api/regularSchedule:
    get:
      summary: 定期スケジュール取得
      operationId: getRegularSchedule
      tags:
        - 定期スケジュール
      parameters:
        - name: from
          in: query
          required: true
          description: 検索開始日（YYYY-MM-DD）
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: 検索終了日（YYYY-MM-DD）
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 正常に取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegularScheduleDTO'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
    post:
      summary: 定期スケジュール登録
      operationId: postRegularSchedule
      tags:
        - 定期スケジュール
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularScheduleInsertForm'
      responses:
        '201':
          description: 登録に成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
    put:
      summary: 定期スケジュール更新
      operationId: putRegularSchedule
      tags:
        - 定期スケジュール
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularScheduleUpdateForm'
      responses:
        '204':
          description: 更新に成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

  /api/regularSchedule/{id}:
    delete:
      summary: 定期スケジュール削除
      operationId: deleteRegularSchedule
      tags:
        - 定期スケジュール
      parameters:
        - name: id
          in: path
          required: true
          description: 削除対象スケジュールID
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: 削除に成功
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

  /api/defaultSchedule:
    get:
      summary: デフォルトスケジュール取得
      operationId: getDefaultSchedule
      tags:
        - デフォルトスケジュール
      parameters:
        - name: from
          in: query
          required: true
          description: 検索開始日（YYYY-MM-DD）
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: 検索終了日（YYYY-MM-DD）
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 正常に取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DefaultScheduleDTO'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
    post:
      summary: デフォルトスケジュール登録
      operationId: postDefaultSchedule
      tags:
        - デフォルトスケジュール
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DefaultScheduleInsertForm'
      responses:
        '201':
          description: 登録に成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'
    put:
      summary: デフォルトスケジュール更新
      operationId: putDefaultSchedule
      tags:
        - デフォルトスケジュール
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DefaultScheduleUpdateForm'
      responses:
        '204':
          description: 更新に成功
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

  /api/defaultSchedule/{id}:
    delete:
      summary: デフォルトスケジュール削除
      operationId: deleteDefaultSchedule
      tags:
        - デフォルトスケジュール
      parameters:
        - name: id
          in: path
          required: true
          description: 削除対象スケジュールID
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: 削除に成功
        '400':
          description: リクエストエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

  /api/organizations:
    get:
      summary: 組織一覧取得
      description: 組織コードと名称の一覧を取得します。
      operationId: getOrganizations
      tags:
        - マスタデータ
      responses:
        '200':
          description: 正常に取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrganizationDTO'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

  /api/workTypes:
    get:
      summary: 勤怠種別一覧取得
      description: 勤怠種別のID・名称・カラーコードを取得します。
      operationId: getWorkTypes
      tags:
        - マスタデータ
      responses:
        '200':
          description: 正常に取得
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkTypeDTO'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthError'

components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID

  schemas:
    LoginRequest:
      type: object
      required: [userId, password]
      properties:
        userId:
          type: string
          description: 認証に使用するユーザーID
          example: "00001"
        password:
          type: string
          format: password
          description: パスワード
          example: "password123"

    LoginSuccess:
      type: object
      required: [message, user]
      properties:
        message:
          type: string
          example: Login successful
        user:
          type: string
          description: 認証されたユーザーID
          example: "00001"

    AuthError:
      type: object
      required: [message]
      properties:
        message:
          type: string
          description: エラーメッセージ
          example: Authentication required

    Error:
      type: object
      description: 業務エラー応答
      required: [message, timestamp]
      properties:
        message:
          type: array
          description: エラーメッセージ一覧
          items:
            type: string
          example:
            - "period: Only one of week or month can be provided"
        timestamp:
          type: string
          format: date-time
          description: エラー発生時刻
          example: "2025-01-01T09:00:00"

    UserDTO:
      type: object
      description: ユーザーとそのスケジュール
      required: [userName, organizationName, schedules]
      properties:
        userName:
          type: string
          example: "山田太郎"
        organizationName:
          type: string
          example: "開発部"
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleDTO'

    ScheduleDTO:
      type: object
      description: 個別スケジュール情報
      required: [scheduleId, date]
      properties:
        scheduleId:
          type: integer
          description: スケジュールID（0の場合は該当なし）
          example: 1
        date:
          type: string
          format: date
          example: "2025-01-06"
        startTime:
          type: string
          format: time
          nullable: true
          example: "09:00:00"
        endTime:
          type: string
          format: time
          nullable: true
          example: "18:00:00"
        worktypeName:
          type: string
          nullable: true
          example: "在宅勤務"
        worktypeColor:
          type: string
          nullable: true
          example: "#2196F3"
        scheduleType:
          type: string
          nullable: true
          enum:
            - REGULAR
            - IRREGULAR
            - DEFAULT
            - COMMON_REGULAR
            - COMMON_IRREGULAR
            - COMMON_DEFAULT

    IrregularScheduleDTO:
      type: object
      description: 非定期スケジュール
      required: [scheduleId, date]
      properties:
        scheduleId:
          type: integer
          example: 10
        date:
          type: string
          format: date
          example: "2025-01-15"
        startTime:
          type: string
          format: time
          nullable: true
          example: "10:00:00"
        endTime:
          type: string
          format: time
          nullable: true
          example: "15:00:00"
        worktypeName:
          type: string
          example: "有給休暇"
        worktypeColor:
          type: string
          example: "#FF9800"

    IrregularScheduleInsertForm:
      type: object
      required: [startTime, endTime, date, workTypeId]
      properties:
        id:
          type: integer
          nullable: true
        startTime:
          type: string
          format: time
          example: "09:00:00"
        endTime:
          type: string
          format: time
          example: "18:00:00"
        date:
          type: string
          format: date
          example: "2025-01-15"
        workTypeId:
          type: string
          pattern: '^(0[1-9]|10|11)$'
          example: "03"

    IrregularScheduleUpdateForm:
      allOf:
        - $ref: '#/components/schemas/IrregularScheduleInsertForm'
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              example: 5

    RegularScheduleDTO:
      type: object
      description: 定期スケジュール
      required: [scheduleId, startDate, endDate, daysOfWeek]
      properties:
        scheduleId:
          type: integer
          example: 3
        startTime:
          type: string
          format: time
          nullable: true
          example: "09:00:00"
        endTime:
          type: string
          format: time
          nullable: true
          example: "17:30:00"
        startDate:
          type: string
          format: date
          example: "2025-01-01"
        endDate:
          type: string
          format: date
          nullable: true
          example: "2025-03-31"
        daysOfWeek:
          type: string
          enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
          example: MONDAY
        worktypeName:
          type: string
          example: "在宅勤務"
        worktypeColor:
          type: string
          example: "#2196F3"

    RegularScheduleInsertForm:
      type: object
      required: [startTime, endTime, startDate, endDate, dayOfWeek, workTypeId]
      properties:
        startTime:
          type: string
          format: time
          example: "09:00:00"
        endTime:
          type: string
          format: time
          example: "18:00:00"
        startDate:
          type: string
          format: date
          example: "2025-01-01"
        endDate:
          type: string
          format: date
          example: "2025-03-31"
        dayOfWeek:
          type: string
          enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
          example: MONDAY
        workTypeId:
          type: string
          pattern: '^(0[1-9]|10|11)$'
          example: "02"

    RegularScheduleUpdateForm:
      allOf:
        - $ref: '#/components/schemas/RegularScheduleInsertForm'
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              example: 8

    DefaultScheduleDTO:
      type: object
      description: デフォルトスケジュール
      required: [scheduleId, startDate, endDate]
      properties:
        scheduleId:
          type: integer
          example: 4
        startTime:
          type: string
          format: time
          nullable: true
          example: "09:00:00"
        endTime:
          type: string
          format: time
          nullable: true
          example: "17:00:00"
        startDate:
          type: string
          format: date
          example: "2025-01-01"
        endDate:
          type: string
          format: date
          nullable: true
          example: "2025-12-31"
        worktypeName:
          type: string
          example: "出社"
        worktypeColor:
          type: string
          example: "#4CAF50"

    DefaultScheduleInsertForm:
      type: object
      required: [startTime, endTime, startDate, endDate, workTypeId]
      properties:
        startTime:
          type: string
          format: time
          example: "09:00:00"
        endTime:
          type: string
          format: time
          example: "18:00:00"
        startDate:
          type: string
          format: date
          example: "2025-01-01"
        endDate:
          type: string
          format: date
          example: "2025-12-31"
        workTypeId:
          type: string
          pattern: '^(0[1-9]|10|11)$'
          example: "01"

    DefaultScheduleUpdateForm:
      allOf:
        - $ref: '#/components/schemas/DefaultScheduleInsertForm'
        - type: object
          required: [id]
          properties:
            id:
              type: integer
              example: 6

    OrganizationDTO:
      type: object
      required: [organizationCode, organizationName]
      properties:
        organizationCode:
          type: string
          example: "006"
        organizationName:
          type: string
          example: "開発部"

    WorkTypeDTO:
      type: object
      required: [id, workTypeName, workTypeColor]
      properties:
        id:
          type: string
          example: "02"
        workTypeName:
          type: string
          example: "在宅勤務"
        workTypeColor:
          type: string
          example: "#2196F3"


tags:
  - name: 認証
    description: 認証・セッション管理
  - name: スケジュール
    description: スケジュール検索
  - name: 非定期スケジュール
    description: 非定期スケジュールの操作
  - name: 定期スケジュール
    description: 定期スケジュールの操作
  - name: デフォルトスケジュール
    description: デフォルトスケジュールの操作
  - name: マスタデータ
    description: 組織・勤怠種別などのマスターデータ取得